split out favourite track row creation
