+++ packages/app/app/components/FavoriteTracksView/styles.scss
@@ -39,7 +39,7 @@
         color: rgba($white, 0.5) !important;
         font-size: 20px !important;
         background: transparent !important;
-        // border-bottom: 0.5px solid rgba($white, 0.2) !important;
+        border-bottom: 1.0px solid rgba($white, 0.2) !important;
       }
 
       i {
